Update Item schema to use integer for Price
Money should be represented as integers vs floats. See the link for more information -> https://husobee.github.io/money/float/2016/09/23/never-use-floats-for-currency.html

+++ lib/rebay_api/listings/item.ex
@@ -7,7 +7,7 @@ defmodule RebayApi.Listings.Item do
     field :description, :string
     field :end_date, :utc_datetime
     field :image, :string
-    field :price, :float
+    field :price, :integer
     field :title, :string
 
     timestamps()
